Create flume2hbase.conf
Configuration file that is used to run a Flume Agent which will read events from a Kafka Producer (via the Kafka Channel) and transport it to HBase (via the HBase sink). This file was wrote on the Hortonworks HDP sandbox and all the path files/configures are linked to my personal sandbox, so will need to be changed for any new implementation of this code.